cli: refac splitting into smaller files
